move notify tests
